Prices working with RXJava. All data inserted except plants and locations.

+++ app/src/main/java/arboretum/arboretumwojslawice/Model/businessentity/Price.java
@@ -46,7 +46,7 @@ public class Price implements AdapterItem {
         return amount;
     }
 
-    public String getAmountString() { return String.valueOf(amount);}
+    public String getAmountString() { return String.valueOf(amount.intValue());}
 
     public void setAmount(Double amount) {
         this.amount = amount;
